+++ include/nodejs/node_osrm_support.hpp
@@ -830,6 +830,9 @@ inline bool parseCommonParameters(const v8::Local<v8::Object> &obj, ParamType &p
         if (annotations->IsBoolean())
         {
             params->annotations = Nan::To<bool>(annotations).FromJust();
+            params->annotations_type = params->annotations
+                                           ? osrm::RouteParameters::AnnotationsType::All
+                                           : osrm::RouteParameters::AnnotationsType::None;
         }
         else if (annotations->IsArray())
         {
@@ -876,6 +879,9 @@ inline bool parseCommonParameters(const v8::Local<v8::Object> &obj, ParamType &p
                     Nan::ThrowError("this 'annotations' param is not supported");
                     return false;
                 }
+
+                params->annotations =
+                    params->annotations_type != osrm::RouteParameters::AnnotationsType::None;
             }
         }
         else